update : applying new logic to /write page
